+++ collect_app/src/main/java/org/odk/collect/android/formlists/savedformlist/SavedFormListViewModel.kt
@@ -1,0 +1,21 @@
+package org.odk.collect.android.formlists.savedformlist
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.ViewModel
+import org.odk.collect.android.instancemanagement.InstancesDataService
+import org.odk.collect.async.Scheduler
+import org.odk.collect.forms.instances.Instance
+
+class SavedFormListViewModel(
+    private val scheduler: Scheduler,
+    private val instancesDataService: InstancesDataService
+) : ViewModel() {
+
+    val formsToDisplay: LiveData<List<Instance>> = instancesDataService.instances
+
+    fun deleteForms(databaseIds: LongArray) {
+        scheduler.immediate(background = true) {
+            databaseIds.forEach { instancesDataService.deleteInstance(it) }
+        }
+    }
+}
