+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyComponent.java
@@ -5,7 +5,7 @@ import android.app.Application;
 import org.javarosa.core.reference.ReferenceManager;
 import org.odk.collect.android.activities.AboutActivity;
 import org.odk.collect.android.activities.AppListActivity;
-import org.odk.collect.android.activities.DeleteSavedFormActivity;
+import org.odk.collect.android.activities.DeleteFormsActivity;
 import org.odk.collect.android.activities.FirstLaunchActivity;
 import org.odk.collect.android.activities.FormDownloadListActivity;
 import org.odk.collect.android.activities.FormFillingActivity;
@@ -211,7 +211,7 @@ public interface AppDependencyComponent {
 
     void inject(ProjectPreferencesFragment projectPreferencesFragment);
 
-    void inject(DeleteSavedFormActivity deleteSavedFormActivity);
+    void inject(DeleteFormsActivity deleteFormsActivity);
 
     void inject(SelectMinimalDialog selectMinimalDialog);
