+++ collect_app/src/main/java/org/odk/collect/android/mainmenu/MainMenuViewModelFactory.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import org.odk.collect.android.application.initialization.AnalyticsInitializer
-import org.odk.collect.android.formmanagement.InstancesDataService
+import org.odk.collect.android.instancemanagement.InstancesDataService
 import org.odk.collect.android.instancemanagement.autosend.AutoSendSettingsProvider
 import org.odk.collect.android.projects.ProjectsDataService
 import org.odk.collect.android.utilities.FormsRepositoryProvider
