+++ collect_app/src/test/java/org/odk/collect/android/formlists/savedformlist/DeleteSavedFormFragmentTest.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.instancemanagement
+package org.odk.collect.android.formlists.savedformlist
 
 import android.content.Context
 import androidx.lifecycle.MutableLiveData
